rename: 일정 목록 조회, 일정 상세, 모임 상세 페이지 및 컴포넌트 폴더명 수정 - 일정 목록 : plans - 일정 상세: plnaDetail - 모임 상세: meetingDetail - 경로도 함께 수정
