+++ src/components/meetingDetail/MeetingDetailFooter.tsx
