+++ src/components/plans/RenderCommonContent.tsx
