--- src/pages/findGatherings/[id].tsx
@@ -1,3 +1,0 @@
-export default function FindGatherings() {
-  return <div>FindGatherings</div>;
-}
